fix: load `createContentScriptMount` for `.ts` files too (not only `.tsx`)
this limitation was, and still is, pretty meh,
because it's implicit, and hard to find.

---

the default template one gets when running `plasmo init`
has `content.ts` (NOT `.tsx`), but includes react content.

this will cause confusion, esp. for new users.

further, a .ts file can import a component from a .tsx file,
thus this restriction for only .tsx files does not make sense.

---

i don't think this is the proper solution,
but will help in the meantime.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>