fix: Allow React to be detected in NPM Workspaces (#864)
* Allow React to be detected in mono repo projects

React is detected by looking for a react devDependency or dependency. NPM Workspaces (mono repos) do not define these in the project package.json. The mono repo will fail if they are defined. The mono repo root has the dependencies listed.

When looking for React in this type of environment, the peerDependencies should be inspected instead.

* fix for react not loading

---------

Co-authored-by: L <6723574+louisgv@users.noreply.github.com>