fix(docs): clarify in the CLI help that argument values must be specified with `=`
Unlike many CLIs, the `plasmo` CLI requires argument values to be
provided with `=` notation (e.g. `--target=chrome-mv3`).

This requirement comes from the `getFlag` method in the
`@plasmo/utils` package.

https://github.com/PlasmoHQ/plasmo-utils/blob/2838ca6214117bfd93b566b4b75586f6a9bd89f7/flags.ts#L8-L16

This expectation isn't clear in the documentation in the CLI, which
is quite misleading. This fixes that, updating all of the examples
to clearly show setting a value with `=`.