fix(docs): document the `--zip` flag in the `plasmo build` CLI (#1051)
Document the `--zip` flag in the `plasmo build` CLI

The `plasmo build` CLI supports a `--zip` flag which builds your
extension and then packages it into a ZIP file.

This is mentioned in the Plasmo docs, but not in the in-CLI
help, which I found confusing.

This adds it to the help output, in the process splitting the
documentation for the `dev` and `build` commands.